main test suite now handles isPartOf relations
