+++ doc/examples/elementclasses-part.xhtml
@@ -4,6 +4,8 @@
        property="dcterms:title"
        content="Title or name of the part"
        class="sectionalelement">
+    <span href="http://example.org/doc"
+	  rel="dcterms:isPartOf" />
     <span about="http://example.org/doc#S1a"
 	  property="dcterms:identifier"
 	  content="Doc pt 1(a)" />
