+++ ferenda/sources/legal/se/sfs.py
@@ -239,7 +239,7 @@ class Register(CompoundElement):
         self.rubrik = kwargs.get('rubrik', None)
         super(Register, self).__init__(*args, **kwargs)
 
-    def as_xhtml(self, uri=None):
+    def as_xhtml(self, uri=None, parent_uri=None):
         res = super(Register, self).as_xhtml()
         res.insert(0, E('h1', self.rubrik))
         return res
@@ -256,7 +256,7 @@ class Registerpost(CompoundElement):
         self.uri = kwargs.get("uri", None)
         super(Registerpost, self).__init__(*args, **kwargs)
 
-    def as_xhtml(self, uri=None):
+    def as_xhtml(self, uri=None, parent_uri=None):
         # FIXME: Render this better (particularly the rpubl:andring
         # property -- should be parsed and linked)
         return super(Registerpost, self).as_xhtml()
