+++ test/testDocRepo.py
@@ -650,346 +650,6 @@ class Repo(RepoTester):
 """)
 
 
-    # class RenderXHTML(RepoTester) # maybe
-    # helper method
-    def _test_render_xhtml(self, body, want):
-        doc = self.repo.make_document('basefile')
-        doc.body = body
-        outfile = self.datadir + "/test.xhtml"
-        self.repo.render_xhtml(doc, outfile)
-        self.assertEqualXML(want, util.readfile(outfile, "rb"))
-        
-    def test_render_xhtml_simple(self):
-        # Test 1: Simple document using our own element objects
-        from ferenda import elements as el
-        body = el.Body([el.Heading(['Toplevel heading'], level=1),
-                        el.Paragraph(['Introductory preamble']),
-                        el.Section([el.Paragraph(['Some text']),
-                                    el.Subsection([el.Paragraph(['More text'])],
-                                                  ordinal='1.1',
-                                                  title="First subsection")],
-                                   ordinal='1', title='First section'),
-                        el.Section([el.Paragraph(['Even more text'])],
-                                   ordinal='2', title='Second section')])
-        want = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">
-    <h1>Toplevel heading</h1>
-    <p>Introductory preamble</p>
-    <div content="First section"
-         about="http://localhost:8000/res/base/basefile#S1"
-         property="dcterms:title"
-         typeof="bibo:DocumentPart"
-         class="section">
-      <span content="1" about="http://localhost:8000/res/base/basefile#S1"
-            property="bibo:chapter"/>
-      <p>Some text</p>
-      <div content="First subsection"
-           about="http://localhost:8000/res/base/basefile#S1.1"
-           property="dcterms:title"
-           typeof="bibo:DocumentPart"
-           class="subsection">
-        <span content="1.1" about="http://localhost:8000/res/base/basefile#S1.1"
-              property="bibo:chapter"/>
-        <p>More text</p>
-      </div>
-    </div>
-    <div content="Second section"
-         about="http://localhost:8000/res/base/basefile#S2"
-         property="dcterms:title"
-         typeof="bibo:DocumentPart"
-         class="section">
-      <span content="2" about="http://localhost:8000/res/base/basefile#S2"
-            property="bibo:chapter"/>
-      <p>Even more text</p>
-    </div>
-  </body>
-</html>"""
-        self._test_render_xhtml(body, want)
-
-    def test_render_xhtml_html(self):
-        # test 2: use element.html elements only, to make a similar
-        # document (although without metadata about
-        # sections/subsection and classses). Uses some HTML5 elements
-        # that are converted to divs when rendering as XHTML 1.1
-        from ferenda.elements import html
-        body = html.Body([html.H1(['Toplevel heading']),
-                          html.Summary(['Introductory preamble']),
-                          html.Section([html.H2(['First section']),
-                                        html.P(['Some text']),
-                                        html.Section([
-                                            html.H3(['First subsection']),
-                                            html.P(['More text'])])]),
-                          html.Section([html.H2(['Second section']),
-                                        html.P(['Even more text'])])])
-        want = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">
-    <h1>Toplevel heading</h1>
-    <div class="summary">Introductory preamble</div>
-    <div class="section">
-      <h2>First section</h2>
-      <p>Some text</p>
-      <div class="section">
-        <h3>First subsection</h3>
-        <p>More text</p>
-      </div>
-    </div>
-    <div class="section">
-      <h2>Second section</h2>
-      <p>Even more text</p>
-    </div>
-  </body>
-</html>
-"""
-        self._test_render_xhtml(body, want)
-
-    def test_render_xhtml_meta(self):
-        from ferenda import elements as el
-        from ferenda.elements import html
-        # test 3: use a mix of our own elements and html elements,
-        # with meta + uri attached to some nodes
-        g1 = rdflib.Graph().parse(format='n3', data="""
-@prefix bibo: <http://purl.org/ontology/bibo/> .
-@prefix dcterms: <http://purl.org/dc/terms/> .
-
-<http://localhost:8000/res/base/basefile#S1> a bibo:DocumentPart;
-        dcterms:title "First section";
-        bibo:chapter "1" .
-        """)
-        g2 = rdflib.Graph().parse(format='n3', data="""
-@prefix bibo: <http://purl.org/ontology/bibo/> .
-@prefix dcterms: <http://purl.org/dc/terms/> .
-@prefix owl: <http://www.w3.org/2002/07/owl#> .
-@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-<http://localhost:8000/res/base/basefile#S2> a bibo:DocumentPart;
-        dcterms:title "Second section";
-        bibo:chapter "2";
-        dcterms:creator "Fred Bloggs"@en-GB;
-        dcterms:issued "2013-05-10"^^xsd:date;
-        owl:sameAs <http://example.org/s2> .
-
-<http://example.org/s2> dcterms:title "Same same but different" .
-       
-<http://localhost:8000/res/base/unlrelated> dcterms:title "Unrelated document" .
-        
-        """)
-        
-        body = el.Body([el.Heading(['Toplevel heading'], level=1),
-                        html.P(['Introductory preamble']),
-                        html.Div([html.P(['Some text']),
-                                  el.Subsection([el.Paragraph(['More text'])],
-                                                ordinal='1.1',
-                                                title="First subsection")],
-                                 uri = 'http://localhost:8000/res/base/basefile#S1',
-                                 meta = g1),
-                        el.Section([el.Paragraph(['Even more text'])],
-                                   uri = 'http://localhost:8000/res/base/basefile#S2',
-                                   meta = g2)])
-        want  = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:owl="http://www.w3.org/2002/07/owl#"
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">
-    <h1>Toplevel heading</h1>
-    <p>Introductory preamble</p>
-    <div about="http://localhost:8000/res/base/basefile#S1"
-         content="First section"
-         property="dcterms:title"
-         typeof="bibo:DocumentPart">
-      <span content="1"
-            property="bibo:chapter"
-            xml:lang=""/>
-      <p>Some text</p>
-      <div about="http://localhost:8000/res/base/basefile#S1.1"
-           content="First subsection"
-           property="dcterms:title"
-           typeof="bibo:DocumentPart"
-           class="subsection">
-        <span about="http://localhost:8000/res/base/basefile#S1.1"
-              content="1.1"
-              property="bibo:chapter"/>
-        <p>More text</p>
-      </div>
-    </div>
-    <div about="http://localhost:8000/res/base/basefile#S2"
-        class="section"
-        content="Second section"
-        property="dcterms:title"
-        typeof="bibo:DocumentPart">
-      <span href="http://example.org/s2"
-            rel="owl:sameAs">
-        <span content="Same same but different"
-              property="dcterms:title"
-              xml:lang=""/>
-      </span>
-      <span content="2"
-            property="bibo:chapter"
-            xml:lang=""/>
-      <span content="2013-05-10"
-            property="dcterms:issued"
-            datatype="xsd:date"/>
-      <span content="Fred Bloggs"
-            property="dcterms:creator"
-            xml:lang="en-GB"/>
-      <p>Even more text</p>
-    </div>
-  </body>
-</html>"""
-        self._test_render_xhtml(body, want)
-
-    def test_render_xhtml_custom(self):
-        # test 4: define a CompoundElement subclass and override
-        # as_xhtml
-        from ferenda import elements as el
-        class Preamble(el.CompoundElement):
-            tagname = "div"
-            classname = "preamble"
-            
-            def as_xhtml(self, uri):
-                # a fairly complicated custom serialization that
-                # inserts a new child node where before there was only
-                # text, and so that text has to be moved from the
-                # parent.text to child.tail
-                E = ElementMaker(namespace="http://www.w3.org/1999/xhtml")
-                element = super(Preamble, self).as_xhtml(uri)
-                note  = E('span', {'class': 'preamble-note'},
-                          self.note + ": ")
-                note.tail = element.text
-                element.text = None
-                element.insert(0, note)
-                return element
-        body = el.Body([el.Heading(['Toplevel heading'], level=1),
-                        Preamble(['Introductory preamble'],
-                                 note='Read this first'),
-                        el.Section([el.Paragraph(['Some text'])],
-                                   ordinal='1', title='First section')])
-
-        want = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">
-    <h1>Toplevel heading</h1>
-    <div class="preamble"><span class="preamble-note">Read this first: </span>Introductory preamble</div>
-    <div content="First section"
-         about="http://localhost:8000/res/base/basefile#S1"
-         property="dcterms:title"
-         typeof="bibo:DocumentPart"
-         class="section">
-      <span content="1" about="http://localhost:8000/res/base/basefile#S1"
-            property="bibo:chapter"/>
-      <p>Some text</p>
-    </div>
-  </body>
-</html>
-"""
-        self._test_render_xhtml(body,want)
-
-
-    def test_render_xhtml_nested(self):
-        # Make sure nested sections have the expected dcterms:isPartOf
-        # relation if they have a @about property (could be from a
-        # .uri property or dynamically constructed like
-        # SectionalElement.as_xhtml)
-        from ferenda import elements as el
-
-        class MySection(el.CompoundElement):
-            partrelation = rdflib.Namespace(util.ns['schema']).isPartOf
-        
-        from ferenda import elements as el
-        body = el.Body([el.Section([el.Paragraph(['Some text']),
-                                    el.Link("txt", uri="http://ex.org/ext"),
-                                    el.Subsection([el.Paragraph(['More text'])],
-                                                  ordinal="1.1",
-                                                  title="First subsection")],
-                                   ordinal="1",
-                                   title="First section"),
-                        MySection([el.Paragraph(['Even more text'])],
-                                   uri="http://example.org/s2")])
-                                   
-        want = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:schema="..."
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">
-    <div about="http://localhost:8000/res/base/basefile#S1"
-         typeof="bibo:DocumentPart"
-         class="section">
-      <span rel="dcterms:isPartOf"
-            href="http://localhost:8000/res/base/basefile"/>
-      <span content="1" about="http://localhost:8000/res/base/basefile#S1"
-            property="bibo:chapter"/>
-      <p>Some text</p>
-      <div content="First subsection"
-           about="http://localhost:8000/res/base/basefile#S1.1"
-           typeof="bibo:DocumentPart"
-           class="subsection">
-        <span rel="dcterms:isPartOf" href="http://localhost:8000/res/base/basefile#S1"/>
-        <span content="1.1" about="http://localhost:8000/res/base/basefile#S1.1"
-              property="bibo:chapter"/>
-        <p>More text</p>
-      </div>
-    </div>
-    <div about="http://localhost:8000/res/base/basefile#S2"
-         typeof="bibo:DocumentPart"
-         class="section">
-      <span rel="schema:isPartOf"
-            href="http://localhost:8000/res/base/basefile"/>
-      <span content="2" about="http://localhost:8000/res/base/basefile#S2"
-            property="bibo:chapter"/>
-      <p>Even more text</p>
-    </div>
-  </body>
-</html>"""
-        # FIXME: uncomment and fix
-        # self._test_render_xhtml(body, want)
-
-    def test_render_xhtml_malformed(self):
-        # Test 5: Illegal indata (raw ESC character in string)
-        from ferenda import elements as el
-        body = el.Body(['Toplevel\x1b heading'])
-        want = """<html xmlns="http://www.w3.org/1999/xhtml"
-                        xmlns:bibo="http://purl.org/ontology/bibo/"
-                        xmlns:dcterms="http://purl.org/dc/terms/"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        version="XHTML+RDFa 1.1"
-                        xsi:schemaLocation="http://www.w3.org/1999/xhtml http://www.w3.org/MarkUp/SCHEMA/xhtml-rdfa-2.xsd"        
-                        xml:lang="en">
-  <head about="http://localhost:8000/res/base/basefile"/>
-  <body about="http://localhost:8000/res/base/basefile">Toplevel heading</body>
-</html>"""
-        self._test_render_xhtml(body, want)
-
-
-
     def test_render_xhtml_head(self):
         doc = self.repo.make_document('basefile')
         headmeta = rdflib.Graph().parse(format='n3', data="""
@@ -1042,42 +702,6 @@ class Repo(RepoTester):
 </html>"""
         self.assertEqualXML(want, util.readfile(outfile, "rb"))
         
-
-    # FIXME: Move this test to a new test case file (testElements.py or even testElementsHtml.py)
-    # class Elements(RepoTester)
-    def test_elements_from_soup(self):
-        from ferenda.elements import html
-        soup = BeautifulSoup("""<body>
-<h1>Sample</h1>
-<div class="main">
-<img src="xyz.png"/>
-<p>Some <b>text</b></p>
-<dl>
-<dt>Term 1</dt>
-<dd>Definition 1</dd>
-</dl>
-</div>
-<div id="foot">
-<hr/>
-<a href="/">home</a> - <a href="/about">about</a>
-</div>
-</body>""")
-        body = html.elements_from_soup(soup.body)
-        # print("Body: \n%s" % serialize(body))
-        result = html.Body([html.H1(["Sample"]),
-                            html.Div([html.Img(src="xyz.png"),
-                                      html.P(["Some ",
-                                              html.B(["text"])]),
-                                      html.DL([html.DT(["Term 1"]),
-                                               html.DD(["Definition 1"])])
-                                ],**{"class":"main"}),
-                            html.Div([html.HR(),
-                                html.A(["home"],href="/"),
-                                " - ",
-                                html.A(["about"],href="/about")
-                        ],id="foot")])
-        self.maxDiff = 4096
-        self.assertEqual(serialize(body),serialize(result))
 
         
     # class Relate(RepoTester)
